Adding null and greater than functions, renaming cond to lesser
I use to null to be able to indicate a default Node with no
type (default 0, but also explicitly set to null which is the first enum
and therefore also 0). This is very handy for the visit() function which
must return a Node reference; if I haven't found the target, I can
return a default "empty" node.

Also added the greater than condition to complement lesser than
condition. For consistency.